chore: auto fixes from pre-commit hooks
# Conflicts:
#	tests/integration/test_design.py
